pyproject.toml
[tool.poetry]
name = "foxops-client"
# managed by poetry-dynamic-versioning, do not change.
version = "0.0.0"
description = "Foxops API Client"
authors = [
    "Alexander Hungenberg <alexander.hungenberg@roche.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
httpx = "^0.24.1"
tenacity = "^8.2.2"
pydantic = "^2.0.3"
structlog = "^23.1.0"

[tool.poetry.group.dev.dependencies]
# Linting
black = "^23.7.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.9.23"
isort = "^5.12.0"

# Testing
pytest = "^7.4.0"
pytest-asyncio = "^0.20.1"
pytest-cov = "^4.1.0"
pytest-docker-tools = "^3.1.3"
pytest-mock = "^3.10.0"

# Typing
mypy = "^1.4.1"
GitPython = "^3.1.29"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[[tool.mypy.overrides]]
module = "pytest_docker_tools"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"
